refact: Merge common and python backend directories, simplifying intermediate layers.
