+++ backend/backend-python/__init__.py
@@ -2,7 +2,7 @@
 
 # pylint: disable=invalid-name  # Module name "backend-python" required by existing structure
 
-from .handler import PythonHandler as Handler
+from .handler import Handler
 from .model_factory import create_model_and_fusion_map
 
 __all__ = ["Handler", "create_model_and_fusion_map"]
