+++ backend/backend-python/adapter_utils.py
