+++ backend/backend-python/model/__init__.py
@@ -1,0 +1,1 @@
+"""Model-related common components."""
