--- backend/common_python/common_model/__init__.py
@@ -1,1 +1,0 @@
-"""Model-related common components."""
